requirements.txt
PyQt6
opencv-python
rembg
Pillow
numpy
onnxruntime
